feat(dashboardnavigation): create dashboard navigation component
